requirements.txt
cachetools

charset-normalizer

google-api-core
google-api-python-client
google-auth
google-auth-httplib2

google-cloud-core
google-cloud-storage
google-cloud-bigquery

# google-crc32c==1.6.0
# google-resumable-media==2.7.2
# googleapis-common-protos==1.66.0
httplib2
# idna==3.10
# proto-plus==1.26.0
# protobuf==5.29.3
# pyasn1==0.6.1
# pyasn1_modules==0.4.1
# pyparsing==3.2.1

certifi
urllib3

dbt-core
dbt-postgres
dbt-bigquery

# google-cloud-storage
# google-cloud-bigquery
# # google-cloud-bigquery[bqstorage,pandas]
pandas 
pandas-gbq
requests
tqdm

# streamlit

dlt[gs,duckdb]
# dlt[]
